requirements.txt

Django==1.8.2
django-appconf==1.0.2
django-compressor==1.5
django-debug-toolbar==1.3.2
django-extensions==1.5.5
django-imagekit==3.2.7
django-widget-tweaks==1.4.1
ipdb==0.10.1
ipython==3.2.0
pilkit==1.1.13
Pillow==3.3.1
pkg-resources==0.0.0
pytz==2016.6.1
six==1.9.0
sqlparse==0.1.15
